+++ packages/nesd/lib/ui/emulator/tools/emulator_tool.dart
@@ -2,6 +2,7 @@ import 'package:collection/collection.dart';
 
 enum EmulatorTool {
   display('Display', contentWidth: _standardToolWidth, minHeight: 470),
+  audio('Audio', contentWidth: _standardToolWidth, minHeight: 240),
   tileViewer('Tile Viewer', contentWidth: _standardToolWidth, minHeight: 480),
   cartridgeInfo(
     'Cartridge Info',
@@ -33,6 +34,7 @@ enum EmulatorTool {
     EmulatorTool.apuDebug ||
     EmulatorTool.executionLog => true,
     EmulatorTool.display ||
+    EmulatorTool.audio ||
     EmulatorTool.tileViewer ||
     EmulatorTool.cartridgeInfo => false,
   };