+++ packages/nesd/lib/ui/settings/settings.g.dart
@@ -16,6 +16,9 @@ _Settings _$SettingsFromJson(Map<String, dynamic> json) => _Settings(
         json['fastForwardSpeed'],
       ) ??
       FastForwardSpeed.x2,
+  turboSpeed:
+      $enumDecodeNullable(_$TurboSpeedEnumMap, json['turboSpeed']) ??
+      TurboSpeed.x1,
   stretch: json['stretch'] as bool? ?? true,
   showBorder: json['showBorder'] as bool? ?? false,
   showDebugOverlay: json['showDebugOverlay'] as bool? ?? false,
@@ -96,6 +99,9 @@ _Settings _$SettingsFromJson(Map<String, dynamic> json) => _Settings(
   crtFilter: json['crtFilter'] == null
       ? const CrtFilterSettings()
       : _crtFilterFromJson(json['crtFilter']),
+  overscan: json['overscan'] == null
+      ? const Overscan()
+      : _overscanFromJson(json['overscan']),
 );
 
 Map<String, dynamic> _$SettingsToJson(_Settings instance) => <String, dynamic>{
@@ -103,6 +109,7 @@ Map<String, dynamic> _$SettingsToJson(_Settings instance) => <String, dynamic>{
   'lowPassFilter': instance.lowPassFilter,
   'swapDutyCycles': instance.swapDutyCycles,
   'fastForwardSpeed': _$FastForwardSpeedEnumMap[instance.fastForwardSpeed]!,
+  'turboSpeed': _$TurboSpeedEnumMap[instance.turboSpeed]!,
   'stretch': instance.stretch,
   'showBorder': instance.showBorder,
   'showDebugOverlay': instance.showDebugOverlay,
@@ -134,6 +141,7 @@ Map<String, dynamic> _$SettingsToJson(_Settings instance) => <String, dynamic>{
       .map((e) => _$VideoFilterEnumMap[e]!)
       .toList(),
   'crtFilter': _crtFilterToJson(instance.crtFilter),
+  'overscan': _overscanToJson(instance.overscan),
 };
 
 const _$FastForwardSpeedEnumMap = {
@@ -141,6 +149,13 @@ const _$FastForwardSpeedEnumMap = {
   FastForwardSpeed.x3: 'x3',
   FastForwardSpeed.x4: 'x4',
   FastForwardSpeed.max: 'max',
+};
+
+const _$TurboSpeedEnumMap = {
+  TurboSpeed.x1: 'x1',
+  TurboSpeed.x2: 'x2',
+  TurboSpeed.x3: 'x3',
+  TurboSpeed.x4: 'x4',
 };
 
 const _$LogLevelEnumMap = {
@@ -190,6 +205,7 @@ const _$VideoFilterEnumMap = {
 
 const _$EmulatorToolEnumMap = {
   EmulatorTool.display: 'display',
+  EmulatorTool.audio: 'audio',
   EmulatorTool.tileViewer: 'tileViewer',
   EmulatorTool.cartridgeInfo: 'cartridgeInfo',
   EmulatorTool.apuDebug: 'apuDebug',
@@ -237,7 +253,7 @@ final class SettingsControllerProvider
 }
 
 String _$settingsControllerHash() =>
-    r'5e4b5ad4e5b9b92540f132d250cbaeaf51f4bc2c';
+    r'31fe3adb93d3d1035f667eb2dd8e3df248ccf498';
 
 abstract class _$SettingsController extends $Notifier<Settings> {
   Settings build();